+++ src/main/java/org/llorllale/youtrack/api/DefaultComments.java
@@ -29,8 +29,6 @@ import org.apache.http.impl.client.HttpClients;
 
 import org.llorllale.youtrack.api.session.Session;
 import org.llorllale.youtrack.api.session.UnauthorizedException;
-import org.llorllale.youtrack.api.util.response.HttpResponseAsResponse;
-import org.llorllale.youtrack.api.util.response.Response;
 
 /**
  * Default implementation of {@link Comments}.
