+++ src/main/java/org/llorllale/youtrack/api/DefaultFields.java
@@ -28,8 +28,6 @@ import org.llorllale.youtrack.api.jaxb.ProjectCustomField;
 import org.llorllale.youtrack.api.jaxb.ProjectCustomFieldRefs;
 import org.llorllale.youtrack.api.session.Session;
 import org.llorllale.youtrack.api.session.UnauthorizedException;
-import org.llorllale.youtrack.api.util.response.HttpResponseAsResponse;
-import org.llorllale.youtrack.api.util.response.Response;
 
 /**
  * Default impl of {@link Fields}.
