+++ src/main/java/org/llorllale/youtrack/api/DefaultIssueTimeTracking.java
@@ -29,8 +29,6 @@ import org.llorllale.youtrack.api.jaxb.WorkItems;
 
 import org.llorllale.youtrack.api.session.Session;
 import org.llorllale.youtrack.api.session.UnauthorizedException;
-import org.llorllale.youtrack.api.util.response.HttpResponseAsResponse;
-import org.llorllale.youtrack.api.util.response.Response;
 
 /**
  * Default implementation of {@link IssueTimeTracking}.
