+++ src/main/java/org/llorllale/youtrack/api/DefaultProjectTimeTracking.java
@@ -26,7 +26,6 @@ import org.apache.http.impl.client.HttpClients;
 import org.llorllale.youtrack.api.jaxb.Settings;
 import org.llorllale.youtrack.api.session.Session;
 import org.llorllale.youtrack.api.session.UnauthorizedException;
-import org.llorllale.youtrack.api.util.response.HttpResponseAsResponse;
 
 /**
  * Default impl of {@link ProjectTimeTracking}.
