+++ src/main/java/org/llorllale/youtrack/api/DefaultProjects.java
@@ -26,7 +26,6 @@ import org.apache.http.impl.client.HttpClients;
 
 import org.llorllale.youtrack.api.session.Session;
 import org.llorllale.youtrack.api.session.UnauthorizedException;
-import org.llorllale.youtrack.api.util.response.HttpResponseAsResponse;
 
 /**
  * Default implementation of {@link Projects}.
