+++ src/main/java/org/llorllale/youtrack/api/DefaultUsersOfProject.java
@@ -25,7 +25,6 @@ import org.apache.http.impl.client.HttpClients;
 
 import org.llorllale.youtrack.api.session.Session;
 import org.llorllale.youtrack.api.session.UnauthorizedException;
-import org.llorllale.youtrack.api.util.response.HttpResponseAsResponse;
 
 /**
  * Default impl of {@link UsersOfProject}.
