+++ src/main/java/org/llorllale/youtrack/api/HttpResponseAsResponse.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.llorllale.youtrack.api.util.response;
+package org.llorllale.youtrack.api;
 
 import java.io.IOException;
 import org.apache.http.HttpResponse;
@@ -33,7 +33,7 @@ import org.llorllale.youtrack.api.session.UnauthorizedException;
  * @author George Aristy (george.aristy@gmail.com)
  * @since 0.1.0
  */
-public final class HttpResponseAsResponse implements Response {
+final class HttpResponseAsResponse implements Response {
   private final Response base;
 
   /**
@@ -41,7 +41,7 @@ public final class HttpResponseAsResponse implements Response {
    * @param httpResponse the {@link HttpResponse} to be adapted
    * @since 0.1.0
    */
-  public HttpResponseAsResponse(HttpResponse httpResponse) {
+  HttpResponseAsResponse(HttpResponse httpResponse) {
     this.base = 
         new UnauthorizedResponse(
             new ForbiddenResponse(
