+++ src/main/java/org/llorllale/youtrack/api/IdentityResponse.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,48 +14,34 @@
  * limitations under the License.
  */
 
-package org.llorllale.youtrack.api.util.response;
+package org.llorllale.youtrack.api;
 
 import java.io.IOException;
 
 import org.apache.http.HttpResponse;
-import org.apache.http.HttpStatus;
-
 import org.llorllale.youtrack.api.session.UnauthorizedException;
-import org.llorllale.youtrack.api.InputStreamAsString;
 
 /**
- * A {@link Response} that throws an {@link IOException} if the server responds with code
- * 500.
+ * Special {@link Response} that does no validation and just returns the {@link HttpResponse}.
  * 
  * @author George Aristy (george.aristy@gmail.com)
  * @since 0.4.0
  */
-public final class InternalServerErrorResponse implements Response {
-  private final Response base;
+final class IdentityResponse implements Response {
+  private final HttpResponse httpResponse;
 
   /**
    * Ctor.
    * 
-   * @param base the next link in the chain
+   * @param httpResponse the http response
    * @since 0.4.0
    */
-  public InternalServerErrorResponse(Response base) {
-    this.base = base;
+  IdentityResponse(HttpResponse httpResponse) {
+    this.httpResponse = httpResponse;
   }
 
   @Override
   public HttpResponse httpResponse() throws IOException, UnauthorizedException {
-    if (this.base.httpResponse().getStatusLine().getStatusCode() 
-        == HttpStatus.SC_INTERNAL_SERVER_ERROR) {
-      throw new IOException(
-          String.format("500 Internal Server error. Payload: %s", 
-              new InputStreamAsString().apply(this.base.httpResponse().getEntity().getContent()
-              )
-          )
-      );
-    }
-
-    return this.base.httpResponse();
+    return this.httpResponse;
   }
 }
