+++ src/main/java/org/llorllale/youtrack/api/InternalServerErrorResponse.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.llorllale.youtrack.api.util.response;
+package org.llorllale.youtrack.api;
 
 import java.io.IOException;
 
@@ -25,38 +25,37 @@ import org.llorllale.youtrack.api.session.UnauthorizedException;
 import org.llorllale.youtrack.api.InputStreamAsString;
 
 /**
- * Throws an {@link IOException} if the server responds with {@code 400 Bad Request}.
- * 
- * <p>Note: 'Bad Request' should never happen :-)</p>
+ * A {@link Response} that throws an {@link IOException} if the server responds with code
+ * 500.
  * 
  * @author George Aristy (george.aristy@gmail.com)
- * @since 0.7.0
+ * @since 0.4.0
  */
-public final class BadRequest implements Response {
-  private final Response response;
+final class InternalServerErrorResponse implements Response {
+  private final Response base;
 
   /**
    * Ctor.
    * 
-   * @param response the next link in the chain
-   * @since 0.7.0
+   * @param base the next link in the chain
+   * @since 0.4.0
    */
-  public BadRequest(Response response) {
-    this.response = response;
+  InternalServerErrorResponse(Response base) {
+    this.base = base;
   }
 
   @Override
   public HttpResponse httpResponse() throws IOException, UnauthorizedException {
-    if (this.response.httpResponse().getStatusLine().getStatusCode() == HttpStatus.SC_BAD_REQUEST) {
+    if (this.base.httpResponse().getStatusLine().getStatusCode() 
+        == HttpStatus.SC_INTERNAL_SERVER_ERROR) {
       throw new IOException(
-          String.format("Server returned 400 Bad Request. Payload: %s",
-              new InputStreamAsString().apply(
-                  this.response.httpResponse().getEntity().getContent()
+          String.format("500 Internal Server error. Payload: %s", 
+              new InputStreamAsString().apply(this.base.httpResponse().getEntity().getContent()
               )
           )
       );
     }
 
-    return this.response.httpResponse();
+    return this.base.httpResponse();
   }
 }
