+++ src/main/java/org/llorllale/youtrack/api/XmlComment.java
@@ -27,7 +27,6 @@ import org.apache.http.impl.client.HttpClients;
 
 import org.llorllale.youtrack.api.session.Session;
 import org.llorllale.youtrack.api.session.UnauthorizedException;
-import org.llorllale.youtrack.api.util.response.HttpResponseAsResponse;
 
 /**
  * JAXB implementation of {@link Comment}.
