--- src/main/java/org/llorllale/youtrack/api/util/package-info.java
@@ -1,26 +1,0 @@
-/* 
- * Copyright 2017 George Aristy (george.aristy@gmail.com).
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-/**
- * <p>
- * The APIs in this package are reserved for internal use by this library.
- * </p>
- *
- * <p>
- * This is a collection of utility classes.
- * </p>
- * @since 0.1.0
- */
-package org.llorllale.youtrack.api.util;
