--- src/main/java/org/llorllale/youtrack/api/util/response/ParseException.java
@@ -1,57 +1,0 @@
-/* 
- * Copyright 2017 George Aristy (george.aristy@gmail.com).
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.llorllale.youtrack.api.util.response;
-
-import java.io.IOException;
-
-/**
- * <p>
- * Signals an error with parsing the payload received from YouTrack.
- * </p>
- * 
- * <p>
- * This is a runtime exception due to the assumption that this kind of error
- * should not be expected.
- * </p>
- * 
- * @author George Aristy (george.aristy@gmail.com)
- * @since 0.1.0
- */
-public class ParseException extends IOException {
-  private static final long serialVersionUID = -8989519985743709400L;
-
-  /**
-   * Ctor.
-   * 
-   * @param message exception message
-   * @since 0.1.0
-   */
-  public ParseException(String message) {
-    super(message);
-  }
-
-  /**
-   * Ctor.
-   * 
-   * @param message exception message
-   * @param cause cause of exception
-   * @since 0.1.0
-   */
-  public ParseException(String message, Throwable cause) {
-    super(message, cause);
-  }
-}
